Tung

+++ src/main/java/com/nutrimeal/nutrimeal/repository/ComboRepository.java
@@ -9,7 +9,7 @@ import java.util.List;
 @Repository
 public interface ComboRepository extends JpaRepository<Combo, Integer> {
 
-    List<Combo> findAllByComboCaloriesBetweenAndIsActiveTrue(Float min, Float max);
+    List<Combo> findAllByComboCaloriesBetweenAndIsActiveTrue(int min, int max);
 
 
     List<Combo> findAllByIsActiveTrue();
